refactor: keep the overlay registration mutators off the stores barrel
acquire/release were public API, so any module could mint a registration with a token it had no way to release — a count that only grows, leaving the Actions menu inert for the rest of the session. They stay exported for the module's own tests but are absent from the barrel, which is the boundary `boundaries/dependencies` actually enforces, so outside `lib/stores` the only way to hold one is useRegisteredOverlay minting its token inside the effect.

Moving the hook in beside the store is what makes that possible; it also removes the need to brand the token, since no caller can supply one.
